Setting impression video details using GAM params
